Update install.sh
Updated installer to list the directories to be deleted.  It also tests saying *nothing* when a users chooses to continue the install.  My logic is that since there's no actual wait at this point, there's no reason to say anything (or to have periods in an animated fashion-- there's no time for them to animate)

Changes per #18